Doc: Avoid duplicate ids in styling macros
The quickcontrols subproject provides and uses a series of QDoc macros,
styleproperty/stylemethod/stylemethod2, that are used to provide a
stylized version of certain documentation element in the output
documentation.

Those macros takes a series of arguments and expand to certain styling
structures that are output format specific.

One of the arguments for each macro, is an identifier that is used in
the expanded structure.

As the macros are format specific, QDoc expands them into "raw" output
for each format, without further processing.

When doing so, the identifier that is provided as an argument, while it
expands to a link-able reference in the format-specific output, will not
be registered by QDoc, which uses an internal linking system, so that
linking to that reference with a QDoc command will fail.

To avoid the issue, each usage of the macro, as documented in the macro
definition, provides a "\target" command with the same identifier, that
QDoc expands to a link-able reference that is registered in the internal
linking system, allowing cross-referencing to work.

Nonetheless, since the same identifier is expanded both by the macro
usage and by the "\target" command adjacent to it, two equivalent
references will appear in the output.

In the DocBook format, on of the possible output format for
QDoc-generated documentation, duplicated references are a violation,
thus producing invalid output.

Since providing a link-able reference to the documented element still
requires the usage of a "\target" command and the expanded to reference
is not essential to the structure provided by the macros, the
styleproperty/stylemethod/stylemethod2 macros are now modified to not
require an identifier argument and avoid expanding to a named reference.

Each usage of the macro was modified to remove the now unused identifier
argument as a consequence of the removal.

Change-Id: Ibf093911ab07ff579d54dc85ed63dbb66bf0114a
Reviewed-by: Thibaut Cuvelier <cuvelier.thibaut@gmail.com>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>