TableView: start a new selection when using Qt::ShiftModifier
If no selection exists in TableView from before, and the
user shift-clicks on a cell, then start a new selection
from the current cell towards the clicked cell. This
is already how it works if a selection is started using
the keyboard, and also standard selection behavior for
other item view implementations outside of Qt (e.g Finder
on macOS).

Fixes: QTBUG-120628
Pick-to: 6.7 6.6 6.5
Change-Id: I2fd34e680eed9992dd7c405fbd58ae4f5b0834ec
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>