QSGRhiTextureGlyphCache: Optimize prepareGlyphImage
In the non-resizeWithTextureCopy case, copy the glyph out of the texture
instead of sending the entire image() (with a source rect set) to avoid
the image detaching and copying the massive image when filling another
one while it's still referenced by QRhiTextureSubresourceUploadDescription.

Furthermore, drop no-op code in ARGB32, taking the r, g, b, a, and then
feeding them back into the pixel makes no difference. qRgb doesn't do
do premultiplication or endianness or anything.

Finally, skip the entire loop for ARGB32 if BGR textures are supported.

Fixes: QTBUG-120007
Pick-to: 6.7 6.6 6.5
Change-Id: I4f712e0a52f0b945d2587103d31189a6c302cc0c
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>