Fix binding loop in FileDialog
The patch 1cc20d181bdee1131bf2eb191e7f8fe4e4927e03 introduced a nested
dialog inside the FileDialog, which contained a Label as its
contentItem. This would cause a binding loop, since the dialog would add
itself as an implicitSize listener to the Label, and attempt to update
it's own implicitSize twice, when the Label's text changes.

Fix the issue by giving the dialog an explicit width.

When the overwriteConfirmationDialog had an animation (which the
Material Dialog has), it was possible to enter
QQuickPopupPositioner::reposition(), when its parentItem (the
fileDialogListView), no longer had a window.

Add a check to return early, in those situations, to avoid a crash.

Pick-to: 6.7 6.5
Fixes: QTBUG-127619
Change-Id: I0324a2a470f237ba1a65a95d8bea7fa50166d756
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 036beb99621781b406b70e6c31841215393e38ea)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>