TableView: start a new selection when using Qt::ShiftModifier (PressAndHold)
Follow up on previous patch to qquickselectionrectangle, and
implement starting a new selection when using Qt::ShiftModifier
also for PressAndHold. This configuration, PressAndHold + holding
Shift to start a selection, is normally not used in combination.
But at the same time, there seems to be no reason why it should't
work either, as using selections handles on desktop is allowed.

Task-number: QTBUG-120628
Pick-to: 6.7 6.6 6.5
Change-Id: I8a7589ffd4d16e339a5654605ee0916d96c8cbf0
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>