DOM refactoring. DomUniverse::parse. Change signature & return type
This commit does the following:
1. Extracts callback call to the callsite of the parse function, where
the "failed" path calls the callback too. It allows to get rid of the
callback argument in the parse function, but requires it to return a
pair of DomItems (arguments for the callback) ->
2. Changes return type of the parse function

This change is helpful because it makes it easier to implement "early
exits" in the parse function and get rid of the god boolean
"skipParse" in the following commits, which will result in better
readability and more straightforward code flow.

Task-number: QTBUG-119550
Change-Id: I0bb9db01561b6eebb95059910bba6fbba80047f4
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>