qv4mm: Make sweep incremental, too
This commit splits the sweep phase into multiple sub-phases. Note that
this is safe, as we already are careful about running user code in sweep
for Component.onDestruction callbacks; for JS objects, we don't execute
arbitrary code in their finalizers, so that's also safe.
The one thing we need to be cautious about is that we don't split the
nulling of the weak values (mainly QObjectWrappers) from actually
freeing the pending QObjectWrapper; otherwise the isUndefined/empty
check could report a wrong result.

Task-number: QTBUG-119274
Change-Id: I57210ac161c897d5946f7c000162d0a30701af19
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>