Improve finding the fillside for horizontal lines
Finding the fillside depends on drawing an imaginary line to x =
infinty and counting the intersections with the polygon. This does
not work reliably if the line is horizontal.

To make this work for horizontal lines, we switch x and y in the
respective algorithm.

Change-Id: I3f625204783c7de2e822bf42d5aaaca6f321dec4
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
(cherry picked from commit 17d94525b122acae735babd88527cf4a100d9af7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>