Fix binding loop and polish issue in quick layout
The quick layout item caused a binding loop issue when layout item sizes
were updated in between polish. This has been fixed by not allowing
rearrange during geometry change, if the layout is already dirty due to
polish event.

There is also a polish issue due to the child item not being invalidated
and it can lead to skipping the corresponding item box size calculation.
This patch invalidates all the items in the rearrange list of the layout
and finally, invalidates the engine and layout.

The threshold limit to trigger layout polish has also been removed now.

Fixes: QTBUG-117899
Fixes: QTBUG-118511
Pick-to: 6.7 6.6 6.5
Change-Id: Ie5713f51ed9e428786046bb06a822e5668beebb0
Reviewed-by: Lars Schmertmann <lars.schmertmann@governikus.de>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>