QQuickSelectionRectangle: only start a new selection if no selection is ongoing
Follow up of aecb62403eb067ac. Ensure we only start a new
selection if no selection is currently active. If you
start a drag while holding down ControlModifier, the first
cell will already be selected by the time the drag handler
becomes active. In that case, we just continue extending
the active selection. We don't actually need to test the
ShiftModifier at this point, since this will be tested
already on press.

Pick-to: 6.7 6.6 6.5
Change-Id: I7af892a495ad62bc540deebd62fae2267cdeb626
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>