DOM refactoring. Generalize updateEntry and move inside DomUniverse
It was not clear why to have updateEntry as a "free" function, passing
Mutex and the corresponding Path map from the Universe. Maybe the
motivation was to have better parallelization, when multiple
threads are competitevely performing multiple updates in a form of
a thread pool. As it was mentioned in the previous commits,
currently DomUniverse is indeed used in multithreaded env, however
the management of the tasks and threads is controlled by another
entity. Hence it makes it a little bit simpler and clear to have
updateEntry as a function member of the DomUniverse. In the end
it's performing updates on the private members of the Universe.
Also in the long run, it would be nice to have only one map for all the
external Item Pairs.
Taking both points into accound this commits attempts to improve
encapsulation and make updateEntry more generic.

Change-Id: Iff38bb6856e261448ddff37b2301dbc6bcabdd4d
Task-number: QTBUG-119550
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>