doc: change names of qmllint and qmlls pages
Make the names of the qmllint and qmlls pages unique by appending
"Reference". This is in-line with what we do in other places of the
documentation, see also the `Qt Creator Manual` page that is used to link
to `Qt Creator`

This fixes the problem of \l{QML Language Server} linking to a
subsection instead to linking to the `QML Language Server Reference`
page.

Task-number: QTBUG-121134
Pick-to: 6.7
Change-Id: I187042d1825d10ea59b5c3af96b8f5fdce10fa9b
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>