Filesystem Explorer: Guard against mobile usage
It was observed that users try to run this example on mobile platforms.
It was never designed for such a usecase and is therefore not supported.

Ref: https://camg.me/qt-mobile-2023/#filesystem

Additionally mention that this is a desktop example in the docs.

Pick-to: 6.6
Change-Id: I55205e3f2cb3a45aaf18f2a52ea24f54ab8e39bd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 9e4f24e7491ebd8e26a7ca4d56046e558b57154c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>