Correct merge reorder issue with the revert patch
This is a follow-up patch to revert the complete change made in patch
9ad9d05f26184a79ad3fe09abdc662949088d19d and
d85de8da6011731e3ff31ae1361e71fecaedb0be.

There are separate patches made to revert these changes but the order in
which these revert patches merged (in 6.7 branch) was incorrect. This
further triggers issue in other branches. This patch fixes the issue
happened due to merge reordering.

Its to be noted that patch d85de8da6011731e3ff31ae1361e71fecaedb0be not
been merged to 6.6 and 6.5 branches but still this patch is required to
revert the complete change in those branches.

Pick-to: 6.6 6.5
Change-Id: I875b13741bccc36c4a4892e5d71b6c7aa7e3b031
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>