doc: explain the visual parent hierarchy of popup better
The visual parent hierarchy of popup is poorly documented.

In most situations, the user only needs to understand that popups appear
in front of everything else in the scene. But there might be use cases
where it's useful to make an item appear in front of a popup.
This can be really tricky, unless you know the internals of popup.

This patch adds another section to the documentation, that hopefully
explains it well enough.

Fixes: QTBUG-118857
Change-Id: If9a638c346951c09a7826df54f590e7328aac764
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 81e53a7e9cf7821c4e512c40671f56ec9b0c52fc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>