qmlls: add Qt Quick completion snippets
Add snippets for Qt Quick types, for feature parity with the Qt Creator
code model.

Do not propose them when QtQuick is not imported, and add the qualifier
to the snippets in case QtQuick is imported with a qualified name, like
`import QtQuick as QQ` for example.

TODO: move the snippets to their own plugin.

Task-number: QTBUG-119969
Change-Id: I874989e28c27d65fcde90b0ccaa4c447c44a5b30
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>