qv4mm: Update non-managed memory limits at correct point in time
With the incremental gc, we might have not freed any memory at all after
the gc call, as we might only be in the marking phase.
Instead, update the limits when we know that we're done, after sweeping
has completed.
Additionally, move the assertion that the gc does not lose memory
there, too, so that it also is called in all sweep calls.

Change-Id: I444ff8817f39153f6f81bb6d21b6b87ae253a5f9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>