Curve renderer: Handle ambiguous points
A triangle could be mis-assigned to the wrong path element if a
point could be found in multiple sub-paths. This change uses a
QMultiHash instead of a QHash to look up the points from the
triangulator, and chooses the path element that has the
triangle on the fill side.

Change-Id: Ic2773da5c7b4a64386e20c19745ce36e2dcb2c57
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 782d9d4e3dccadeec3ef58a64b38e2a3426e9bd6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>