QQmlEngine: Set translationsDirectory in loadFromModule
In the process, loadFromModule had to be split in two parts. The type
resolution has to happen so that the translationDirectory can be set and
the translations have to be loaded before loading the actual module.

Pick-to: 6.5
Fixes: QTBUG-116589
Change-Id: Ife7999f418ba35bfb0eed9050198e5a886fa74ae
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 986b4d155bd39836b1aa3c3da332ef07d2a949fd)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>