String::append:  Avoid growing list more than necessary
To visit all nodes, we should need at most depth of the rope many
entries. So far, we might however have added more nodes, as we
immediately pushed both the left and the right child. By storing a
marker together with the node, we can avoid populating the list with
both the left and right subtree at the same time.

Change-Id: Ife095c439bb20a7cf5546c31918d19869263ece1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>