Refactor svgtoqml tool
Move the generation part of the tool to a separate private module,
which can generate either a QML file or a QQuickItem representation
of an SVG File. The generator relies mainly on QtQuickShapes for
rendering, in addition to using other QQuickItems.

The generator in this module can be used by 'svgtoqml' tool. The plan
succeeding this commit will be to add a QML module that will expose a
QML component for loading an SVG file at runtime.

Task-number: QTBUG-121659
Change-Id: I4014eb696a7d39f05f8b3035bdcee50d2acaf2b6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>