QQuickMenu: sync action with native menu item
Make sure we update the text of a native menu
item when the text of the corresponding
QQuickActions changes.

Since the argumens between the signal and the
slot differs, we need to use Q_PRIVATE_SLOT
syntax instead of PointerToMember. Using
lambdas is also inconvenient, since we
need to be able to disconnect the signal again
upon removing the action.

Change-Id: I41421232c7215183600212c07447e353db07c832
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>