qmlls: enable qualified type completion in Bindings
Amends 29d2153f8adeceed56df432ccb775ff28a8e3347: this commit added
support for qualified types to reachableTypes, but forgot to adapt
insideBindingCompletion, a method that uses reachableTypes, to the new
functionality for qualified types.

Add a test for qualified types in bindings, and rename the
'asCompletion' test to qualifiedTypeCompletionAfterDot.

Task-number: QTBUG-120111
Pick-to: 6.7
Change-Id: I5fe9718e65310503e998d71ee4da7f18edcf688c
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>