QQuickMenuPrivate: add a nativeHandle() function
Add a nativeHandle() function that can be called
from e.g QQuickMenuBar. Since this function has the
"side effect" that it creates the native menu if it
doesn't exists, add a maybeNativeHandle() as well
that can be used when we only need a pointer the
native menu if it already exists, but can otherwise
be skipped (such as when removing menu items).

Keep the existing usingNativeMenu(), since this, in
comparison to nativeHandle(), will only create a native
menu when a native menu is requested. The name of the
function also makes the code more readable from where
it's used.

Change-Id: I9bd15f0bed25a4af7744dac5184fb4f597bdc269
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>