Compiler: Don't crash on optional lookups on QJSValue base type
Handle the case where the base type of the lookup is QJSValue so that we
don't hit the assert.

However, it should not be possible to get a QJSValue there at all. This
should be investigated further. Created QTBUG-121662.

Fixes: QTBUG-121393
Change-Id: I8bea87cbff74119bb977635ec391601f47269ff0
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 3672a173df7a518ff375dcd5931a1ce5255d4115)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>