Fix a bug in intersection solving
When we hit an intersection, we use the angle between the segments to
decide in which direction to continue. This concept falls apart when we
intersect a path at the join between two segments. For this case we have
to fall back to check the fill side. This is more expensive and thus not
done for every intersection.

Change-Id: Iebefc84fe3d299e5c31817ba34b3a1cdcb1e7d68
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit c1208af5004d0f200b471bfe8f3b396be7940edd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>