Replace TextEdit.textDocument.error signal with status property
If we make loading and saving asynchronous later, we'll need the same
sort of status property as in Image, Loader etc. to let the user QML
know when it's done, or when it has failed. Bindings are more flexible
than signal handlers; so we might as well have better-prepared API now.

To make use of the Status enum in QML, we need to de-anonymize
TextDocument and make it merely uncreatable. Anyway, it was already
referred to as the TextDocument type in the docs.

Change-Id: Ia4354feba77995759797030382891308b941f0af
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 5647b6900b2ecc291022143176b545b933eca3a8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>