Unconditionally create delegates when using native menus
Initially I was trying to avoid creating MenuItem delegates
when using a native menu, but there are several reasons why this is not
ideal:
- It makes managing indices of contentModel vs nativeItems complicated
  and error-prone.
- It complicates the code.
- It introduces differences between native and non-native execution for
  the same app's codebase. Users relying on code to be executed in e.g.
  Component.onCompleted of Menu's delegate would see that that code
  isn't executed when using a native menu.

The disadvantage of the approach in this patch is that we're creating
full MenuItem instances that will never be rendered. Although far from
ideal, this is unlikely to have a noticeable impact on performance on
most platforms where we support native menus.

Task-number: QTBUG-69558
Change-Id: Ibd0c01654ff8ee40087d5472cf9f9233dea78bf3
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>