QQuickMenuBar: ensure we check that a native menubar was successfully created
QPlatformTheme can choose to return nullptr when calling
createPlatformMenuBar(). As such, we need to check that
we actually have a valid pointer afterwards, before using
it.

This patch will therefore implement the same pattern as
in QQuickMenu: add a function usingNativeMenuBar()
that can be called up-front before trying to access the
handle.

Change-Id: I8d0b4c055184332c688280d8bb363c66936afd7a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>