qmlls: make resolveExpressionType use findDefiningScope
It seems that findDefiningScope()'s functionality is not completely
exhaustive, which leads to some bugs when trying to find usages from
different files. Therefore, this patch fixes them to be able to work on
find usages for multiple files.

The find usages functionality compare findDefiningScope()'s result
to find out if two usages are about the same types, and one of the
problems is that for anything that is not an enum, a method or a
property, findDefiningScope() returns a nullptr: and two nullptr are
equal and therefore being reported as an usage.

To fix this, findDefiningScope() needs to be implemented for all kind of
possible QML constructs, and that is best done directly inside of
resolveExpressionType(). Therefore, adapt the ResolveOwnerType mode
in resolveExpressionType() to deal with calling the correct version of
findDefiningScope().

Replace all calls of findDefiningScope() with
resolveExpressionType(..., ResolveDefiningType). Note that this avoids
the nullptr situation described above because the output of
resolveExpressionType() is usually checked against nullptr, in
comparison with the previous usages of findDefiningScope().

Add switch cases in all helper methods of resolveExpressionType() to
take into account all QQmllSUtilsResolveOptions.

The findDefiningScope method was split in multiple parts, such that
it can be used to check for scopes defining a method, a property,
etc... findDefiningScope can't be used to check everything at once
because then you end up with "collisions", for example scopes with
property bindings might clash with scopes with property definitions,
and depending on the caller you might want the one or the other.

Once the switch-cases were added all over resolveExpressionType and its
helpers, extract some duplicated code into static helpers instead of
fixing each duplicated code on its own.

Task-number: QTBUG-114990
Change-Id: I15accf56c14e53aba6d8357b5e90d9ebcf036ef1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>