QQuickMenuBar: add createNativeMenuBar() and removeNativeMenuBar()
Currently, the native menubar is (partially) created as a side-effect
of calling nativeHandle(). Since calling this function can be done
from many places (also from outside of QQuickMenuBar), it gives an
unpredictable construction phase that can change as a result of
code changing outside of QQuickMenuBar.

To gain a better and more deterministic control over when, and
how, we create and delete the native menubar, this patch will
add three new functions: createNativeMenuBar(),
removeNativeMenuBar(), and syncNativeMenuBarVisible().
Using them will simplify the code, and make the visibility
state of the native menubar easier to debug and maintain.

Change-Id: Ifaf9783a2edffccdbfc94ed7688581d971e92d83
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>