Remove nested paths during intersection solving
Nested paths that were neither generating a new hole or a new filling
resulted in double filling of some areas. The previous workaround
was to force the intersectionSolver to generate a new path, however,
which is computational expensive and does not allow to check for
changes.

This patch introduces a dedicated function to remove superfluos
nested subpaths. It should be faster than the previous solution and
allows to track changes to the path.

With this patch, the intersectionSolver returns true only if the path
was changed.

Pick-to: 6.7
Change-Id: I16919f20b4e5460ddd729e8be7703fb2879c3190
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>