QQuickMenuBarPrivate: add functions insertMenu and takeMenu
There are many ways to add menus to a menuBar; You
can declare them as children explicitly from QML,
or indirect as children of MenuBarItems, or dynamically
by callng addMenu or insertMenu.

Funnel all ways of adding and removing menus into
QQuickMenuBarPrivate::insertMenu() and
QQuickMenuBarPrivate::takeMenu() to ensure that we
implement the same logic regardless of how the menus
are added and removed.

This change detected some memory management bugs in
the addRemove auto test, where a QScopedPointer was
used to store a menuBar resolved from a QML file. This
should not be done, as the menubar is owned by the
QML engine, and should not be destroyed by the
scoped pointer when it goes out of scope. The same
with the menus; they are owned by the menubar, and
should not be destroyed from the outside without
calling "takeMenu()" first (which the test already
checks).

Change-Id: I3f818c3d76dd7c3fa9e1eaa3025597f015b69ce8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>