Compiler: Assert against noop instructions
The left-over instructions of QTBUG-109261 were only able to persist
because the dead code analysis assumes that instructions aren't noops
and because the StoreElement instruction was not marked as having side
effects. Since a521216b7149185de9c82b9b6b34486dac8794a1, StoreElement
has been marked as having side effects and the left-over instructions
can thus no longer remain. The dead code elimination logic is thus fine
as it is.

Add an assert in QQmlJSTypePropagator::endInstruction() to prevent noop
instructions in the future.

Fixes: QTBUG-109261
Pick-to: 6.7 6.6 6.5
Change-Id: Ida1a121f016c17954ac0005690c100b468378198
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>