PersistentValue: fix gc isues
Heap objects shouldn't be put on the markStack, but instead be marked
(so that their own black bit is set, too).
Also, consistenyl call set in all copy ctors and assignement operators
to ensure that marking happens there, too. The compiler should be able
to optimize the superfluous null checks out in the ctors.

Change-Id: I45df06d1ba569d61467455a7e4628162abd1bd9f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>