sg: Fix some material-related multiview issues
Every time type is compared and compare() is called, viewCount() must
also be compared.

More importantly, the correct view count (handled via internal flags
for SC/BC) has to be pushed to the QSGMaterial always, before checking
for an existing QSGMaterialShader - otherwise if material #2 has a hit
for a shader from material #1, material #2 will never have the view
count pushed so it continues to report viewCount() == 1, which
has...interesting consequences. Avoid this.

Change-Id: I0808cc57f53a8dab891d1b36b41790f58c978ef4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>