QQuickMenuBar: don't delete menus that might still be referenced from QML
The current implementation documents that it "Removes and
destroys the specified menu" from removeMenu(). This is not
safe. Consider this code:

Menu { id: menu }
MenuBar {
   id: menuBar
   Component.onCompleted: addMenu(menu)
}
Text { text: menu.title }

If the app after a while calls menuBar.removeMenu(menu), we
would be deleting 'menu' from c++. This would cause it to
be dangling, and perhaps cause a crash when the text binding
is later reevaluated.

The idea behind the MenuBar taking over memory management
for the Menus and MenuBarItems, was probably to cater for
the situation where you use the Component API to dynamically
create menus that you feed to addMenu(). But even in that
case the developer doesn't need to explicitly destroy anything;
The gc will eventually come around and do the job.

This patch will therefore change this, so that we never delete
any menus using deleteLater, but leave it to the gc to find
the right time. This is now also clarified in the docs.

Change-Id: I654874f8b79fc90ccb818e688ed73183dace6c32
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>