sg: Allow adopting foreign textures with GL_SRGB8_ALPHA8
Do what the other (Vulkan, D3D) paths do: set the appropriate
QRhiTexture flag while mapping to RGBA8. This has no visible effect,
but it avoids flooding the debug output with warnings in Quick 3D XR
applications on Android (Quest 3) when using OpenGL ES.
(because it hits the 'default' case that prints a warning and
continues with UnknownFormat; not having the correct enum value
in QRhiTexture has no effect in typical Qt Quick / 3D apps, but
nonetheless we should not hit that path)

The usage of sRGB formats is an interesting case with VR compositors:
with Vulkan and D3D12 we explicitly drop the _SRGB format in
Quick 3D XR (e.g. we change VK_FORMAT_R8G8B8A8_SRGB to
VK_FORMAT_R8G8B8A8_UNORM) before passing the native image
object to Qt Quick, but there this is necessary since having a
_SRGB format changes the behavior in shader writes (because the
image/render target views inherit the QRhiTexture format
in the QRhi backends), whereas Qt Quick 3D performs
linearization and sRGB conversion at the end of the pipeline,
so this would clash (like doing an additional unwanted
linear-sRGB conversion on the already sRGB data)

OpenGL is different, however, when it comes to the API: there is
no implicit behavior unless the appropriate state is enabled
(which we do not do), hence the OpenGL (ES) path in Quick 3D XR
does not need to perform the format demotion, but then it is
ideal if Qt Quick accepts those formats (which was missing
until now for some reason unknown). Regardless, the demotion
may still be implemented in Quick 3D XR separately later.

In practice this should remove the
"GL format 35907 is not supported"
messages appearing in the Android logs on every frame or so
with Quick 3D XR apps.

Change-Id: I3e9135f790be0eda51d3b1d68f6d115eb1a21000
Reviewed-by: Andy Nichols <andy.nichols@qt.io>