Curve renderer: fix crash when path ends with moveTo
If a quad path ends with a moveTo, possibly because a very short line
has been optimized away, the subPathsClosed() function would copy the
state that says that a new sub-path is to start. This would in turn
confuse the lineTo() function that is used to close the path. The result
was that the final element would be a start element instead of an end
element, which breaks the invariant that every start element has an end.
This causes a crash when the invalid path is sent to the intersection
solver.

Change-Id: I89617b74f23e59354a4a5031412fad60d97ec887
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 32b5ee7842b8042e3d805209d13a34bb873ee280)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>