ArrayData/MemberData: Prove Fixed gc interaction
When setting values on a newly allocated Array- or MemberData,
d7aa952e143accc18d54707d956d019272197078 made the assumption that it is
safe to skip the write-barrier, as the new values would be marked when
the "values" member would be written to (pushing the new data on the
mark stack, and then later marking all objects stored in values when it
gets popped and markObjects runs).
Now that we no longer do black allocations, this actually holds true.
Add a unit test to verify it.

Task-number: QTBUG-119274
Task-number: QTBUG-121910
Change-Id: Ia1ceeaffeaf30dc1fb2b9e1992dd0b599050294c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>