QQuickMenuBar: don't set parent item when incubating a MenuBarItem
When we incubate a MenuBarItem, we currently set the parent item
already from QQuickMenuBarPrivate::beginCreateItem(). This is
unfortunate, since doing so will end up in a call to
QQuickContainerPrivate::itemChildAdded() before the MenuBarItem
has finished incubating, which is not safe. Moreover, this will
also add the new MenuBarItem to the back of all the other existing
container items, which is not what we want if the construction
was done as a result of calling MenuBar.insertMenu(index, menu).

Setting the parent item this early is anyway not needed, since
QQuickContainer will make sure to set the correct parent item
in the follow up call to QQuickContainer::insertItem().

So, just remove the explicit setting of the parent item from
the beginCreateItem. This will make sure that calls such
as MenuBar.insertMenu(0, menu) will work.

Change-Id: Ie2b759cfdd1b7a15be0ff646d6d9c6bf0c4c8328
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>